Melhorias usabilidade
