+++ lib/widgets/util/navigation_utils.dart
@@ -25,6 +25,10 @@ class NavigationUtils {
     navigateTo(context, OperacaoPage(tipoOperacao, ativo));
   }
 
+  static void navigateToAtivosCarteira( BuildContext context) async {
+    navigateTo(context, AtivosCarteiraPage());
+  }
+
   static Future<dynamic> navigateTo( BuildContext context, Widget widget){
     return Navigator.push(context, MaterialPageRoute(builder: (context) {
       return widget;
@@ -54,7 +58,6 @@ class NavigationUtils {
   static  Future<dynamic>  replaceWithOperacao( BuildContext context, TipoOperacao tipoOperacao, Ativo ativo) async {
     return replacePageWith(context, OperacaoPage(tipoOperacao, ativo));
   }
-
 
 
   static void showMessage(final BuildContext context, String text, {Function onVisible} ) {
@@ -126,4 +129,53 @@ class NavigationUtils {
       return widget;
     }));
   }
+
+
+  static void showMenuOperacaoAtivo(BuildContext context, Ativo ativo){
+    showModalBottomSheet<void>(
+        context: context,
+        builder: (BuildContext context) {
+          return Container(
+              child: Wrap(children: <Widget>[
+                _menuComprarAtivo(context, ativo),
+                _menuVenderAtivo(context, ativo)
+
+              ]
+              )
+          );
+        });
+  }
+
+  static Widget _menuComprarAtivo(BuildContext context, Ativo ativo) {
+    var operacao = TipoOperacao.COMPRA;
+    var icon = Icons.add_shopping_cart;
+    var text = 'Comprar ${ativo.ticker}';
+    return _menuItem(context, operacao, ativo, icon, text);
+  }
+
+  static Widget _menuVenderAtivo(BuildContext context, Ativo ativo) {
+    var operacao = TipoOperacao.VENDA;
+    var icon = Icons.remove_shopping_cart;
+    var text = 'Vender ${ativo.ticker}';
+    return _menuItem(context, operacao, ativo, icon, text);
+  }
+
+
+  static ListTile _menuItem(BuildContext context, TipoOperacao operacao, Ativo ativo, IconData icon, String text) {
+    return ListTile(
+        onTap: () {
+          NavigationUtils.close(context);
+          NavigationUtils.navigateToOperacao(
+              context, operacao, ativo);
+        },
+        leading: Icon(
+          icon,
+          size: 30,
+        ),
+        title: Text(text,
+            textAlign: TextAlign.left,
+            style: TextStyle(fontSize: 18.0, fontWeight: FontWeight.bold, color: kNighSky))
+    );
+  }
+
 }
